Update code example to match other .NET code example structure
