--- dotnetv3/Bedrock/Actions/Bedrock/ListFoundationModelsExample/ListFoundationModelsExample/ListFoundationModelsExample.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.7.34202.233
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ListFoundationModelsExample", "ListFoundationModelsExample.csproj", "{51A011A9-2AEA-433D-A6B0-611216F955DF}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{51A011A9-2AEA-433D-A6B0-611216F955DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{51A011A9-2AEA-433D-A6B0-611216F955DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{51A011A9-2AEA-433D-A6B0-611216F955DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{51A011A9-2AEA-433D-A6B0-611216F955DF}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {86D7A4D8-43F7-42C9-AFE9-F8AB12A91A0B}
-	EndGlobalSection
-EndGlobal
