+++ dotnetv3/Bedrock/Actions/HelloBedrock.cs
@@ -1,0 +1,71 @@
+﻿// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved. 
+// SPDX-License-Identifier:  Apache-2.0
+
+// snippet-start:[Bedrock.dotnetv3.HelloBedrock]
+using Amazon;
+using Amazon.Bedrock;
+using Amazon.Bedrock.Model;
+
+namespace ListFoundationModelsExample
+{
+    /// <summary>
+    /// This example shows how to list foundation models.
+    /// </summary>
+    internal class HelloBedrock
+    {
+        /// <summary>
+        /// Main method to call the ListFoundationModelsAsync method.
+        /// </summary>
+        /// <param name="args"> The command line arguments. </param>
+        static async Task Main(string[] args)
+        {
+            // Specify a region endpoint where Amazon Bedrock is available. For a list of supported region see https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html#bedrock-regions
+            AmazonBedrockClient bedrockClient = new(RegionEndpoint.USWest2);
+
+            await ListFoundationModelsAsync(bedrockClient);
+
+        }
+
+        /// <summary>
+        /// List foundation models.
+        /// </summary>
+        /// <param name="bedrockClient"> The Amazon Bedrock client. </param>
+        private static async Task ListFoundationModelsAsync(AmazonBedrockClient bedrockClient)
+        {
+            Console.WriteLine("List foundation models with no filter");
+
+            try
+            {
+                ListFoundationModelsResponse response = await bedrockClient.ListFoundationModelsAsync(new ListFoundationModelsRequest()
+                {
+                });
+
+                if (response?.HttpStatusCode == System.Net.HttpStatusCode.OK)
+                {
+                    foreach (var fm in response.ModelSummaries)
+                    {
+                        WriteToConsole(fm);
+                    }
+                }
+                else
+                {
+                    Console.WriteLine("Something wrong happened");
+                }
+            }
+            catch(AmazonBedrockException e)
+            {
+                Console.WriteLine(e.Message);
+            }
+        }
+
+        /// <summary>
+        /// Write the foundation model summary to console.
+        /// </summary>
+        /// <param name="foundationModel"> The foundation model summary to write to console. </param>
+        private static void WriteToConsole(FoundationModelSummary foundationModel)
+        {
+            Console.WriteLine($"{foundationModel.ModelId}, Customization: {String.Join(", ", foundationModel.CustomizationsSupported)}, Stream: {foundationModel.ResponseStreamingSupported}, Input: {String.Join(", ", foundationModel.InputModalities)}, Output: {String.Join(", ", foundationModel.OutputModalities)}");
+        }
+    }
+}
+// snippet-end:[Bedrock.dotnetv3.HelloBedrock]
